Fixup 'id' in 'LoanTypeRepository'.
* Fixed 'id' property in 'LoanTypeRepository' to Long, instead of Integer - due to 'LoanType' 'id' field type.

+++ loans/src/main/java/com/askie01/loans/repository/LoanTypeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LoanTypeRepository extends JpaRepository<LoanType, Integer> {
+public interface LoanTypeRepository extends JpaRepository<LoanType, Long> {
     Optional<LoanType> findByName(String name);
 }
